Split the jhipster app exporter
It had two responsibilities: formatting & writing
Formatting has been moved to another file while the exporting still does the
two (but only calling the formatter this time)
